# Improve performance of the core function