@dcousens 2022-10-29
Similar strategy to https://github.com/JedWatson/classnames/pull/197, but without all the other changes.  Nice pull request @fromaline.

If I can reproduce this in a number of different, modern environments, then I see no reason why we can't.
The problem has always been that this "improvement" is variable between environments.