@dcousens 2022-11-03
> Do you need help with benchmarking this “improvement” in different environments?
Yes

How to manifest that help at the moment, I'm undecided.
Maybe we could integrate these measurements into CI, or maybe you can at least verify on your end that the results are consistent across what browsers you have available?

In any event, the help is appreciated :pray: