@fromaline 2022-11-03
Ok, I'll think about it, try different approaches and let you know once done.