@fromaline 2022-11-03
I went with the simplest solution, hope it's enough for now. Current benchmarks don't work for me, so I created new ones and extend existing cases. I run all tests on M1 MacBook Air M1, 2020 (8 GB, macOS 12.4). You may find my benchmarks [here](https://github.com/fromaline/classnames/tree/benchmarks).

## The results:
As you said, this "improvement" is variable between environments, so I'm not sure now if we need it at all. 

### Chrome

```
Case strings:
- array-based cn is 4,014,452.027 op/sec
- string-based cn is 4,278,990.158 op/sec
Conclusion: string-based cn is faster by 6.18%

Case object:
- array-based cn is 5,621,135.469 op/sec
- string-based cn is 5,425,935.974 op/sec
Conclusion: array-based cn is faster by 3.47%

Case strings & objects:
- array-based cn is 4,985,044.865 op/sec
- string-based cn is 4,743,833.017 op/sec
Conclusion: array-based cn is faster by 4.84%

Case arrays:
- array-based cn is 2,382,086.708 op/sec
- string-based cn is 2,857,959.417 op/sec
Conclusion: string-based cn is faster by 16.65%

Case strings & objects & arrays:
- array-based cn is 3,121,098.627 op/sec
- string-based cn is 3,434,065.934 op/sec
Conclusion: string-based cn is faster by 9.11%
```

### Safari

```
Case strings:
- array-based cn is 2,512,562.814 op/sec
- string-based cn is 2,506,265.664 op/sec
Conclusion: array-based cn is faster by 0.25%

Case object:
- array-based cn is 2,610,966.057 op/sec
- string-based cn is 2,673,796.791 op/sec
Conclusion: string-based cn is faster by 2.35%

Case strings & objects:
- array-based cn is 2,570,694.087 op/sec
- string-based cn is 2,525,252.525 op/sec
Conclusion: array-based cn is faster by 1.77%

Case arrays:
- array-based cn is 1,082,251.082 op/sec
- string-based cn is 1,132,502.831 op/sec
Conclusion: string-based cn is faster by 4.44%

Case strings & objects & arrays:
- array-based cn is 1,618,122.977 op/sec
- string-based cn is 1,631,321.37 op/sec
Conclusion: string-based cn is faster by 0.81%
```

### Firefox

```
Case strings:
- array-based cn is 3,257,328.99 op/sec
- string-based cn is 2,493,765.586 op/sec
Conclusion: array-based cn is faster by 23.44%

Case object:
- array-based cn is 2,840,909.091 op/sec
- string-based cn is 2,040,816.327 op/sec
Conclusion: array-based cn is faster by 28.16%

Case strings & objects:
- array-based cn is 2,881,844.38 op/sec
- string-based cn is 2,057,613.169 op/sec
Conclusion: array-based cn is faster by 28.60%

Case arrays:
- array-based cn is 2,169,197.397 op/sec
- string-based cn is 1,560,062.402 op/sec
Conclusion: array-based cn is faster by 28.08%

Case strings & objects & arrays:
- array-based cn is 1,626,016.26 op/sec
- string-based cn is 1,340,482.574 op/sec
Conclusion: array-based cn is faster by 17.56%
```

### Node.js

```
Case strings:
- array-based cn is 3,467,621.808 op/sec
- string-based cn is 3,949,787.021 op/sec
Conclusion: string-based cn is faster by 12.21%

Case object:
- array-based cn is 4,705,921.108 op/sec
- string-based cn is 5,016,186.607 op/sec
Conclusion: string-based cn is faster by 6.19%

Case strings & objects:
- array-based cn is 4,113,779.608 op/sec
- string-based cn is 4,570,430.034 op/sec
Conclusion: string-based cn is faster by 9.99%

Case arrays:
- array-based cn is 1,960,371.898 op/sec
- string-based cn is 2,788,989.259 op/sec
Conclusion: string-based cn is faster by 29.71%

Case strings & objects & arrays:
- array-based cn is 2,556,558.534 op/sec
- string-based cn is 3,259,376.282 op/sec
Conclusion: string-based cn is faster by 21.56%
```

### Deno

```
Case strings:
- array-based cn is 3,649,635.036 op/sec
- string-based cn is 3,649,635.036 op/sec
Conclusion: string-based cn is faster by 0.00%

Case object:
- array-based cn is 4,950,495.05 op/sec
- string-based cn is 4,672,897.196 op/sec
Conclusion: array-based cn is faster by 5.61%

Case strings & objects:
- array-based cn is 4,629,629.63 op/sec
- string-based cn is 4,166,666.667 op/sec
Conclusion: array-based cn is faster by 10.00%

Case arrays:
- array-based cn is 2,109,704.641 op/sec
- string-based cn is 2,380,952.381 op/sec
Conclusion: string-based cn is faster by 11.39%

Case strings & objects & arrays:
- array-based cn is 2,673,796.791 op/sec
- string-based cn is 2,824,858.757 op/sec
Conclusion: string-based cn is faster by 5.35%
```