@fromaline 2022-11-04
Oh, do you want to merge my benchmarks?
I can polish them and merge into main branch in forked repo first, if needed.