@dcousens 2022-11-04
> Current benchmarks don't work for me

Out of interest, what went wrong?