@fromaline 2022-11-04
browserify didn’t work for some reason
maybe I skipped some important step, ‘cause I never used it before